requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.10.0
asyncmy==0.2.10
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
bm25s==0.3.2.post1
build==1.3.0
cachetools==6.2.1
cbor==1.0.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.7
dataclasses-json==0.6.7
datasets==4.8.3
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dill==0.4.1
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
filelock==3.20.0
filetype==1.2.0
FlagEmbedding==1.3.5
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.42.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.71.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hashids==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
ijson==3.5.0
ImageIO==2.37.3
img2pdf==0.6.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
inscriptis==2.7.1
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.30.0
lazy-loader==0.5
llama-index-core==0.14.6
llama-index-embeddings-huggingface==0.6.1
llama-index-instrumentation==0.4.2
llama-index-llms-gemini==0.6.2
llama-index-llms-ollama==0.7.4
llama-index-llms-openai==0.7.4
llama-index-llms-openai-like==0.7.1
llama-index-postprocessor-flag-embedding-reranker==0.5.0
llama-index-readers-file==0.5.4
llama-index-retrievers-bm25==0.7.1
llama-index-storage-chat-store-redis==0.5.1
llama-index-vector-stores-lancedb==0.5.0
llama-index-workflows==2.10.0
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.19
mypy_extensions==1.1.0
nanoid==2.0.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numpy==2.4.3
oauthlib==3.3.1
ocrmypdf==16.12.0
ollama==0.6.0
onnxruntime==1.23.2
openai==2.8.1
opencv-python-headless==4.13.0.92
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20251107
peft==0.18.1
pikepdf==10.0.2
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pylance==3.0.1
PyMuPDF==1.26.6
pyparsing==3.3.2
pypdf==6.1.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyStemmer==2.2.0.3
pythainlp==5.3.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
redis==7.0.0b1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scalar_fastapi==1.5.0
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==3.4.1
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
striprtf==0.0.26
sympy==1.14.0
tantivy==0.25.1
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
trec-car-tools==2.6
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zlib-state==0.1.10
